this migrates more test specific functionality to utils simplifying the gating attributions
