+++ zingolib/src/testutils.rs
@@ -11,6 +11,7 @@ use pepper_sync::keys::decode_address;
 use zcash_address::unified::Fvk;
 use zcash_keys::address::UnifiedAddress;
 use zcash_keys::encoding::AddressCodec;
+use zcash_primitives::consensus::NetworkConstants;
 use zcash_protocol::consensus::BlockHeight;
 use zcash_protocol::{PoolType, ShieldedProtocol, consensus};
 use zingo_infra_services::LocalNet;
@@ -620,6 +621,56 @@ pub fn int_to_pooltype(int: i32) -> PoolType {
 /// if someone figures out how to improve this code it can be done in one place right here.
 pub(crate) fn timestamped_test_log(text: &str) {
     println!("{}: {}", crate::utils::now(), text);
+}
+
+#[allow(unused_macros)]
+macro_rules! build_method {
+    ($name:ident, $localtype:ty) => {
+        #[doc = "Set the $name field of the builder."]
+        pub fn $name(&mut self, $name: $localtype) -> &mut Self {
+            self.$name = Some($name);
+            self
+        }
+    };
+}
+
+#[allow(unused_imports)]
+pub(crate) use build_method;
+
+#[allow(unused_macros)]
+macro_rules! build_method_push {
+    ($name:ident, $localtype:ty) => {
+        #[doc = "Push a $ty to the builder."]
+        pub fn $name(&mut self, $name: $localtype) -> &mut Self {
+            self.$name.push($name);
+            self
+        }
+    };
+}
+#[allow(unused_macros)]
+macro_rules! build_push_list {
+    ($name:ident, $builder:ident, $struct:ident) => {
+        for i in &$builder.$name {
+            $struct.$name.push(i.build());
+        }
+    };
+}
+
+#[allow(unused_imports)]
+pub(crate) use build_method_push;
+#[allow(unused_imports)]
+pub(crate) use build_push_list;
+
+/// Take a P2PKH taddr and interpret it as a tex addr
+pub fn interpret_taddr_as_tex_addr(
+    taddr_bytes: [u8; 20],
+    p: &impl zcash_primitives::consensus::Parameters,
+) -> String {
+    bech32::encode::<bech32::Bech32m>(
+        bech32::Hrp::parse_unchecked(p.network_type().hrp_tex_address()),
+        &taddr_bytes,
+    )
+    .unwrap()
 }
 
 /// Decodes unified address and re-encode as sapling address.
